select fix attempt
