+++ src/core/Select/Select.stories.mdx
@@ -51,7 +51,7 @@ Buttons are used when the user is to perform some form of action.
 ### Multiselect
 <Preview>
     <Story name="Multiselect">
-        <Select open multiSelect>
+        <Select multiSelect>
             <Option>Option 1</Option>
             <Option>Option 2</Option>
             <Option>Option 3</Option>
